Rename data URI downloader module.
