+++ scrapy/core/downloader/handlers/datauri.py
